+++ CHANGELOG.md
@@ -5,10 +5,15 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Fixed
+- `make deps`: fix CUDA detection and allow CPU as fallback
+
 ### Changed
 - instance segmentation postprocessing: use asymmetric overlap
   criterion for non-maximum suppression
 - skip instances which belong to classes with empty category
+- annotate incrementally (by skipping candidates that overlap
+  with pre-existing top-level regions)
 
 ## [0.1.0] - 2022-01-21